#187 re-create waypoints when switch maps

+++ Source/core/window/BaseWindow.lua
@@ -59,9 +59,14 @@ function BaseWindow:setShowing(doShow)
 end
 
 function BaseWindow:destroy()
+    if self == nil or not WindowApi.doesExist(self.id) then
+        return
+    end
+
     if self.eventRegister ~= nil then
         self.eventRegister:unregisterAllEvents()
     end
+
     WindowUtils.SaveWindowPosition(self.id)
     WindowApi.destroyWindow(self.id)
     self = nil
